Unit tests so far
Having a bit of trouble with testing the behaviour tree right now. To be
able to run the tests you'll need to make a new native unit test project
in the tank war solution, and move the UnitTest2 folder one folder back.
It's also needed to add a reference to the tankwar project and set it up
for release mode, the unit test project just needs the SFML directories
set up